refactor: widgets 폴더를 shared 폴더 하위로 이동
