+++ lib/features/route_planning/widgets/poi_settings_grid.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:ridingmate/core/design/typography/app_text_style.dart';
 import 'package:ridingmate/core/theme/extensions.dart';
 import 'package:ridingmate/core/theme/semantic_colors.dart';
-import 'package:ridingmate/widgets/button/icon_button_solid.dart';
-import 'package:ridingmate/widgets/icon/icon_size.dart';
+import 'package:ridingmate/shared/widgets/button/icon_button_solid.dart';
+import 'package:ridingmate/shared/widgets/icon/icon_size.dart';
 
 class POISetting {
   const POISetting({required this.icon, required this.label});
