+++ lib/shared/widgets/button/button_solid.dart
@@ -1,26 +1,26 @@
 import 'package:flutter/material.dart';
-import 'package:ridingmate/widgets/button/base_button.dart';
+import 'package:ridingmate/shared/widgets/button/base_button.dart';
 
-class ButtonOutlined extends BaseButton {
-  const ButtonOutlined({
+class ButtonSolid extends BaseButton {
+  const ButtonSolid({
     super.key,
     super.text,
     super.leftIcon,
     super.rightIcon,
     super.size,
     required super.textColor,
-    required this.borderColor,
+    required this.backgroundColor,
     super.onPressed,
     super.shadow,
   });
 
-  final Color borderColor;
+  final Color backgroundColor;
 
   @override
   BoxDecoration getDecoration(BuildContext context) {
     return BoxDecoration(
       borderRadius: BorderRadius.circular(getBorderRadius()),
-      border: Border.all(color: borderColor),
+      color: backgroundColor,
       boxShadow: shadow,
     );
   }
