+++ lib/shared/widgets/input/search_field_preview.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:ridingmate/core/design/decorations/app_shadows.dart';
 import 'package:ridingmate/core/theme/extensions.dart';
 import 'package:ridingmate/core/theme/semantic_colors.dart';
-import 'package:ridingmate/widgets/input/base_search_field.dart';
-import 'package:ridingmate/widgets/input/search_field_size.dart';
+import 'package:ridingmate/shared/widgets/input/base_search_field.dart';
+import 'package:ridingmate/shared/widgets/input/search_field_size.dart';
 
 class SearchFieldPreview extends StatelessWidget {
   const SearchFieldPreview({
