+++ lib/shared/widgets/navigation/search_navigation_bar.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:ridingmate/core/theme/extensions.dart';
 import 'package:ridingmate/core/theme/semantic_colors.dart';
-import 'package:ridingmate/widgets/input/search_field.dart';
-import 'package:ridingmate/widgets/input/search_field_size.dart';
+import 'package:ridingmate/shared/widgets/input/search_field.dart';
+import 'package:ridingmate/shared/widgets/input/search_field_size.dart';
 
 class SearchNavigationBar extends StatelessWidget
     implements PreferredSizeWidget {
