+++ lib/shared/widgets/thumbnail/thumbnail.dart
